requirements.txt
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
cachetools==5.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.4
docutils==0.18.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.15.3
Flask==2.1.2
fonttools==4.33.3
gitdb==4.0.9
GitPython==3.1.27
htmlmin==0.1.12
idna==3.3
ImageHash==4.2.1
importlib-metadata==4.11.3
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
Kivy==2.1.0
kivy-deps.angle==0.3.2
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.4.5
Kivy-Garden==0.1.5
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
missingno==0.5.1
mistune==0.8.4
multimethod==1.8
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8
notebook==6.4.11
numpy==1.22.3
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
pandas-profiling==3.2.0
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
phik==0.12.2
pickleshare==0.7.5
Pillow==9.1.0
plotly==5.8.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
pure-eval==0.2.2
pyarrow==8.0.0
pycparser==2.21
pydantic==1.9.0
pydeck==0.7.1
Pygments==2.12.0
Pympler==1.0.1
pyparsing==3.0.9
pypiwin32==223
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
pywin32==304
pywinpty==2.0.5
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
scikit-learn==1.1.0
scipy==1.8.0
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.2.0
statsmodels==0.13.2
streamlit==1.9.0
streamlit-lottie==0.0.3
streamlit-option-menu==0.3.2
streamlit-pandas-profiling==0.1.3
tangled-up-in-unicode==0.2.0
tenacity==8.0.1
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
toolz==0.11.2
tornado==6.1
tqdm==4.64.0
traitlets==5.2.0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
validators==0.19.0
visions==0.7.4
watchdog==2.1.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.0
xgboost==1.6.1
zipp==3.8.0
